apps/edge-hub/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bidict==0.23.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.122.0
h11==0.16.0
idna==3.11
numpy==2.2.6
opencv-python==4.12.0.88
paho-mqtt==2.1.0
pydantic==2.12.4
pydantic_core==2.41.5
python-engineio==4.12.3
python-socketio==5.15.0
requests==2.32.5
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wsproto==1.3.2
